requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
cloudinary==1.37.0
Pillow==10.2.0
Werkzeug==3.0.1
dnspython==2.6.1
gunicorn==21.2.0
